+++ Library/Homebrew/dev-cmd/bump-revision.rb
@@ -52,9 +52,9 @@ module Homebrew
         Utils::Inreplace.inreplace(formula.path) do |s|
           s = s.inreplace_string
           if current_revision.zero?
-            Utils::AST.add_formula_stanza!(s, name: :revision, text: text)
+            Utils::AST.add_formula_stanza!(s, :revision, text)
           else
-            Utils::AST.replace_formula_stanza!(s, name: :revision, replacement: text)
+            Utils::AST.replace_formula_stanza!(s, :revision, text)
           end
         end
       end
